+++ src/Order/LineItemComposer.php
@@ -7,6 +7,7 @@ namespace Shopgate\Shopware\Order;
 use Shopgate\Shopware\Order\Mapping\LineItem\LineItemProductMapping;
 use Shopgate\Shopware\Order\Mapping\LineItem\LineItemPromoMapping;
 use Shopgate\Shopware\Shopgate\Extended\ExtendedCart;
+use Shopgate\Shopware\System\Log\LoggerInterface;
 use ShopgateCartBase;
 use ShopgateExternalCoupon;
 use Shopware\Core\Checkout\Cart\Cart;
@@ -21,15 +22,22 @@ class LineItemComposer
     private $productMapping;
     /** @var LineItemPromoMapping */
     private $promoMapping;
+    /** @var LoggerInterface */
+    private $logger;
 
     /**
      * @param LineItemProductMapping $productMapping
      * @param LineItemPromoMapping $promoMapping
+     * @param LoggerInterface $logger
      */
-    public function __construct(LineItemProductMapping $productMapping, LineItemPromoMapping $promoMapping)
-    {
+    public function __construct(
+        LineItemProductMapping $productMapping,
+        LineItemPromoMapping $promoMapping,
+        LoggerInterface $logger
+    ) {
         $this->productMapping = $productMapping;
         $this->promoMapping = $promoMapping;
+        $this->logger = $logger;
     }
 
     /**
@@ -58,12 +66,14 @@ class LineItemComposer
          */
         foreach ($cart->getLineItems() as $id => $lineItem) {
             if (!$this->isValidUuid($id)) {
+                $this->logger->debug('Invalid line item id provided: ' . $id);
                 break;
             }
             switch ($lineItem->getType()) {
                 case LineItem::PRODUCT_LINE_ITEM_TYPE:
                     $incomingItem = $sgCart->findItemById($id);
                     if (null === $incomingItem) {
+                        $this->logger->debug('Cannot locate line item in sg cart: ' . $id);
                         break;
                     }
                     $lineItems[$id] = $this->productMapping->mapValidProduct(
@@ -76,9 +86,13 @@ class LineItemComposer
                     $refId = $lineItem->getReferencedId();
                     $sgPromoItem = $sgCart->findExternalCoupon($refId) ?? new ShopgateExternalCoupon();
                     $sgPromoItem->setCode(empty($refId) ? $id : $refId); // for cart_rule
+                    $sgPromoItem->setInternalInfo(empty($refId) ? LineItemPromoMapping::RULE_ID : '');
                     $sgPromoItem->setCurrency($sgCart->getCurrency());
                     $externalCoupons[$id] = $this->promoMapping->mapValidCoupon($lineItem, $sgPromoItem);
                     break;
+                default:
+                    $this->logger->debug('Cannot map item type: ' . $lineItem->getType());
+                    $this->logger->debug(print_r($lineItem->jsonSerialize(), true));
             }
         }
 
@@ -106,6 +120,9 @@ class LineItemComposer
                     $missingCoupon->setNotValidMessage($error->getMessage());
                     $externalCoupons[$error->getParameters()['code']] = $missingCoupon;
                     break;
+                default:
+                    $this->logger->debug('Unmapped cart errors & notifications');
+                    $this->logger->debug(print_r($error->jsonSerialize(), true));
             }
         }