+++ src/System/Di/Forwarder.php
@@ -6,6 +6,7 @@ namespace Shopgate\Shopware\System\Di;
 
 use Shopgate\Shopware\ExportService;
 use Shopgate\Shopware\ImportService;
+use Shopgate\Shopware\System\Log\LoggerInterface;
 
 /**
  * Forwarder for Plugin class where we cannot use
@@ -18,15 +19,18 @@ class Forwarder
     private $exportService;
     /** @var ImportService */
     private $importService;
+    /** @var LoggerInterface */
+    private $logger;
 
     /**
      * @param ExportService $exportService
      * @param ImportService $importService
      */
-    public function __construct(ExportService $exportService, ImportService $importService)
+    public function __construct(ExportService $exportService, ImportService $importService, LoggerInterface $logger)
     {
         $this->exportService = $exportService;
         $this->importService = $importService;
+        $this->logger = $logger;
     }
 
     /**
@@ -43,5 +47,13 @@ class Forwarder
     public function getImportService(): ImportService
     {
         return $this->importService;
+    }
+
+    /**
+     * @return LoggerInterface
+     */
+    public function getLogger(): LoggerInterface
+    {
+        return $this->logger;
     }
 }